bump typenum and include srtool

+++ runtime/metaverse/src/lib.rs
@@ -197,7 +197,7 @@ parameter_types! {
 		.avg_block_initialization(AVERAGE_ON_INITIALIZE_RATIO)
 		.build_or_panic();
 
-	pub const SS58Prefix: u8 = 268;
+	pub const SS58Prefix: u16 = 268;
 }
 
 // Filter call that we don't enable before governance launch
